Back/requirements.txt
attrs==23.1.0
certifi==2023.5.7
charset-normalizer==3.2.0
click==8.1.6
colorama==0.4.6
connexion==2.14.2
Flask==2.2.5
idna==3.4
inflection==0.5.1
itsdangerous==2.1.2
Jinja2==3.1.2
jsonschema==4.18.4
jsonschema-specifications==2023.7.1
MarkupSafe==2.1.3
mysql-connector-python==8.1.0
python-dotenv==1.0.0
PyYAML==6.0.1
requests==2.31.0
Werkzeug==2.2.3
uuid==1.30
bcrypt==4.0.1
moviepy==1.0.3
Flask-Cors==4.0.0
PyJWT==2.8.0